code/requirements.txt
streamlit==1.37.1
numpy==2.0.1
pillow==10.4.0
matplotlib==3.9.2
